+++ src/components/avatar-group/index.ts
@@ -1,2 +1,0 @@
-export { MPAvatarGroup } from './MPAvatarGroup';
-export type { MPAvatarGroupProps } from './MPAvatarGroup';